fix repeated execution of same process bug
In the case of extra signals, especially those at the beginning, the same process was repeatedly executed even though it wasn't its turn. This change ensures that once a process starts it will set toberun_pid to invalid value so that it won't restart over and over again!
I'm sorry if those sounds to complex  to comprehend but, tbh, it is maximum clarity that a person who did not understand the situation can achive at all.
"Di wê çolê winda me/ Nizaim ex b'xwe kî me"